🔨 Changed the location of mcserver-run
Given that the script is only meant to be run by the service and not
all users, it didn't make sense to put it in the PATH. It was removed
from it.

The script itself now only asks for 1G of RAM instead of 5G, as the
system may not have that much memory.

Also, the compress.sh script now has an option to compress using zstd,
and threaded xz compression is now done trough 'xz -T0' instead of
'pxz'.